moved krt library to app layer
